# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: RingiFlow API
  description: ワークフロー管理システム RingiFlow の BFF API
  license:
    name: MIT
    identifier: MIT
  version: 0.1.0
paths:
...
  /api/v1/workflow-definitions:
    get:
      tags:
      - workflows
      summary: GET /api/v1/workflow-definitions
      description: |-
        ワークフロー定義一覧を取得する

        ## 処理フロー

        1. セッションから `tenant_id` を取得
        2. Core Service の `GET /internal/workflow-definitions` を呼び出し
        3. レスポンスを返す
      operationId: list_workflow_definitions
      responses:
        '200':
          description: ワークフロー定義一覧
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkflowDefinitionData'
        '401':
          description: 認証エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
      - session_auth: []
    post:
      tags:
      - workflow-definitions
      summary: POST /api/v1/workflow-definitions
      description: ワークフロー定義を新規作成する（Draft 状態）。
      operationId: create_definition
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDefinitionRequest'
        required: true
      responses:
        '201':
          description: 定義作成成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowDefinitionData'
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
      - session_auth: []
  /api/v1/workflow-definitions/validate:
    post:
      tags:
      - workflow-definitions
      summary: POST /api/v1/workflow-definitions/validate
      description: ワークフロー定義 JSON のバリデーションのみ実行する。保存は行わない。
      operationId: validate_definition
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidateDefinitionRequest'
        required: true
      responses:
        '200':
          description: バリデーション結果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationResultData'
        '400':
          description: リクエスト不正
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
      - session_auth: []
  /api/v1/workflow-definitions/{id}:
    get:
      tags:
      - workflows
      summary: GET /api/v1/workflow-definitions/{id}
      description: |-
        ワークフロー定義の詳細を取得する

        ## 処理フロー

        1. セッションから `tenant_id` を取得
        2. Core Service の `GET /internal/workflow-definitions/{id}` を呼び出し
        3. レスポンスを返す
      operationId: get_workflow_definition
      parameters:
      - name: id
        in: path
        description: ワークフロー定義 ID
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: ワークフロー定義詳細
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowDefinitionData'
        '404':
          description: 定義が見つからない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
      - session_auth: []
    put:
      tags:
      - workflow-definitions
      summary: PUT /api/v1/workflow-definitions/{id}
      description: ワークフロー定義を更新する（Draft のみ）。
      operationId: update_definition
      parameters:
      - name: id
        in: path
        description: ワークフロー定義 ID
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDefinitionRequest'
        required: true
      responses:
        '200':
          description: 定義更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowDefinitionData'
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: 定義が見つからない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '409':
          description: バージョン競合
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
      - session_auth: []
    delete:
      tags:
      - workflow-definitions
      summary: DELETE /api/v1/workflow-definitions/{id}
      description: ワークフロー定義を削除する（Draft のみ）。
      operationId: delete_definition
      parameters:
      - name: id
        in: path
        description: ワークフロー定義 ID
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: 削除成功
        '400':
          description: Draft 以外の削除
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: 定義が見つからない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
      - session_auth: []
  /api/v1/workflow-definitions/{id}/archive:
    post:
      tags:
      - workflow-definitions
      summary: POST /api/v1/workflow-definitions/{id}/archive
      description: ワークフロー定義をアーカイブする（Published → Archived）。
      operationId: archive_definition
      parameters:
      - name: id
        in: path
        description: ワークフロー定義 ID
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublishArchiveRequest'
        required: true
      responses:
        '200':
          description: アーカイブ成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowDefinitionData'
        '400':
          description: Published 以外
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: 定義が見つからない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '409':
          description: バージョン競合
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
      - session_auth: []
  /api/v1/workflow-definitions/{id}/publish:
    post:
      tags:
      - workflow-definitions
      summary: POST /api/v1/workflow-definitions/{id}/publish
      description: ワークフロー定義を公開する（Draft → Published）。
      operationId: publish_definition
      parameters:
      - name: id
        in: path
        description: ワークフロー定義 ID
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublishArchiveRequest'
        required: true
      responses:
        '200':
          description: 公開成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowDefinitionData'
        '400':
          description: バリデーション失敗 or Draft 以外
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: 定義が見つからない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '409':
          description: バージョン競合
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
      - session_auth: []
...
components:
  schemas:
...
    CreateDefinitionRequest:
      type: object
      description: 定義作成リクエスト（BFF 公開 API）
      required:
      - name
      - definition
      properties:
        name:
          type: string
          description: ワークフロー定義名
        description:
          type:
          - string
          - 'null'
          description: 説明（任意）
        definition:
          description: 定義 JSON
...
    ProblemDetails:
      type: object
      description: |-
        エラーレスポンス（RFC 9457 Problem Details）

        すべてのサービスで統一されたエラーレスポンス形式。
        `type` フィールドは URI で問題の種類を識別する。
      required:
      - type
      - title
      - status
      - detail
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
          format: int32
          minimum: 0
        detail:
          type: string
    PublishArchiveRequest:
      type: object
      description: 公開/アーカイブリクエスト（BFF 公開 API）
      required:
      - version
      properties:
        version:
          type: integer
          format: int32
          description: 楽観的ロック用バージョン
...
    UpdateDefinitionRequest:
      type: object
      description: 定義更新リクエスト（BFF 公開 API）
      required:
      - name
      - definition
      - version
      properties:
        name:
          type: string
          description: ワークフロー定義名
        description:
          type:
          - string
          - 'null'
          description: 説明（任意）
        definition:
          description: 定義 JSON
        version:
          type: integer
          format: int32
          description: 楽観的ロック用バージョン
...
    ValidateDefinitionRequest:
      type: object
      description: バリデーションリクエスト（BFF 公開 API）
      required:
      - definition
      properties:
        definition:
          description: 検証対象の定義 JSON
    ValidationErrorData:
      type: object
      description: バリデーションエラーデータ
      required:
      - code
      - message
      properties:
        code:
          type: string
        message:
          type: string
        step_id:
          type:
          - string
          - 'null'
    ValidationResultData:
      type: object
      description: バリデーション結果データ
      required:
      - valid
      - errors
      properties:
        valid:
          type: boolean
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ValidationErrorData'
...
    WorkflowDefinitionData:
      type: object
      description: ワークフロー定義データ
      required:
      - id
      - name
      - version
      - definition
      - status
      - created_by
      - created_at
      - updated_at
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type:
          - string
          - 'null'
        version:
          type: integer
          format: int32
        definition: {}
        status:
          type: string
        created_by:
          type: string
        created_at:
          type: string
        updated_at:
          type: string
...
  securitySchemes:
    session_auth:
      type: apiKey
      in: cookie
      name: session_id
...